Split roda.rb into multiple files
From my perspective, there are several advantages to keeping the base
functionality all inside lib/roda.rb:

* the base functionality can be just another plugin
* you have everything you need inside a single file (like Sinatra)

However, I feel that the base functionality is now large enough to
benefit from splitting into multiple files. In my opinion it makes code
navigation easier as it allows one to focus only on a single class at a
time.

Therefore, we move classes into separate files:

  Roda               => lib/roda.rb (remains)
  Roda::RodaRequest  => lib/roda/request.rb
  Roda::RodaResponse => lib/roda/response.rb
  Roda::RodaPlugins  => lib/roda/plugins.rb
  Roda::RodaCache    => lib/roda/cache.rb
